Address feedback and re-enable non-Linux
