+++ libr/debug/p/native/linux/linux_debug.c
@@ -8,6 +8,7 @@
 #include <signal.h>
 #include <sys/uio.h>
 #include "linux_debug.h"
+#include "../procfs.h"
 
 const char *linux_reg_profile (RDebug *dbg) {
 #if __arm__
@@ -105,30 +106,6 @@ int linux_attach (RDebug *dbg, int pid) {
 	return pid;
 }
 
-static int linux_get_proc_pid(int pid, char *prop, char *out, size_t len) {
-	char filename[128];
-	int fd, ret = -1;
-	ssize_t nr;
-
-	snprintf (filename, sizeof(filename) - 1, "/proc/%d/%s", pid, prop);
-	fd = open (filename, O_RDONLY);
-	if (fd == -1)
-		return -1;
-	nr = read (fd, out, len);
-	if (nr > 0) {
-		out[nr - 1] = '\0';  /* terminate at newline */
-		ret = 0;
-	}
-	else if (nr < 0) {
-		r_sys_perror ("read");
-	}
-	else {
-		eprintf ("linux_get_proc_pid: got EOF reading from \"%s\"\n", filename);
-	}
-	close (fd);
-	return ret;
-}
-
 RDebugInfo *linux_info (RDebug *dbg, const char *arg) {
 	char procpid_cmdline[1024];
 	RDebugInfo *rdi = R_NEW0 (RDebugInfo);
@@ -144,84 +121,6 @@ RDebugInfo *linux_info (RDebug *dbg, const char *arg) {
 		"/proc/%d/cmdline", rdi->pid);
 	rdi->cmdline = r_file_slurp (procpid_cmdline, NULL);
 	return rdi;
-}
-
-#undef MAXPID
-#define MAXPID 99999
-
-RList *linux_pid_list (int pid, RList *list)
-{
-	int i;
-	char *ptr, buf[1024];
-
-	list->free = (RListFree)&r_debug_pid_free;
-	if (pid) {
-		/* add the requested pid. should we do this? we don't even know if it's valid still.. */
-		r_list_append (list, r_debug_pid_new ("(current)", pid, 's', 0));
-
-		/* list parents */
-		DIR *dh;
-		struct dirent *de;
-		dh = opendir ("/proc");
-		if (dh == NULL) {
-			r_sys_perror ("opendir /proc");
-			r_list_free (list);
-			return NULL;
-		}
-		while ((de = readdir (dh))) {
-			/* for each existing pid file... */
-			i = atoi (de->d_name);
-			if (!i)
-				continue;
-
-			/* try to read the status */
-			if (linux_get_proc_pid (i, "status", buf, sizeof(buf)) == -1)
-				continue;
-
-			/* look for the parent process id */
-			ptr = strstr (buf, "PPid:");
-			if (ptr) {
-				int ppid = atoi (ptr + 6);
-
-				/* if this is the requested process... */
-				if (i == pid) {
-					//eprintf ("PPid: %d\n", ppid);
-					/* append it to the list with parent */
-					r_list_append (list, r_debug_pid_new (
-						"(ppid)", ppid, 's', 0));
-				}
-
-				/* ignore it if it is not one of our children */
-				if (ppid != pid)
-					continue;
-
-				/* it's a child of the requested pid, read it's command line and add it */
-				if (linux_get_proc_pid (ppid, "cmdline", buf, sizeof(buf)) == -1)
-					continue;
-
-				r_list_append (list, r_debug_pid_new (buf, i, 's', 0));
-			}
-		}
-		closedir (dh);
-	} else {
-		/* try to bruteforce the processes
-		 * XXX(jjd): wouldn't listing the processes like before work better?
-		 */
-		for (i = 2; i < MAXPID; i++) {
-			/* try to send signal 0, if it fails it must not be valid */
-			if (r_sandbox_kill (i, 0) == -1)
-				continue;
-
-			// TODO: Use slurp!
-			// XXX(jjd): we probably need better zero-copy w/r_list if we use
-			// slurp here
-			if (linux_get_proc_pid(i, "cmdline", buf, sizeof(buf)) == -1)
-				continue;
-
-			r_list_append (list, r_debug_pid_new (buf, i, 's', 0));
-		}
-	}
-	return list;
 }
 
 RList *linux_thread_list (int pid, RList *list) {
@@ -242,8 +141,10 @@ RList *linux_thread_list (int pid, RList *list) {
 		while ((de = readdir (dh))) {
 			int tid = atoi (de->d_name);
 
-			if (linux_get_proc_pid(tid, "comm", buf, sizeof(buf)) == -1)
-				continue;
+			if (procfs_pid_slurp (tid, "comm", buf, sizeof(buf)) == -1) {
+				/* fall back to auto-id */
+				snprintf (buf, sizeof(buf), "thread_%d", thid++);
+			}
 
 			// TODO: get status, pc, etc..
 			r_list_append (list, r_debug_pid_new (buf, tid, 's', 0));
@@ -255,7 +156,7 @@ RList *linux_thread_list (int pid, RList *list) {
 #define MAXPID 99999
 		/* otherwise, brute force the pids */
 		for (i = pid; i < MAXPID; i++) { // XXX
-			if (linux_get_proc_pid(i, "status", buf, sizeof(buf)) == -1)
+			if (procfs_pid_slurp (i, "status", buf, sizeof(buf)) == -1)
 				continue;
 
 			/* look for a thread group id */
@@ -267,7 +168,11 @@ RList *linux_thread_list (int pid, RList *list) {
 				if (tgid != pid)
 					continue;
 
-				snprintf (buf, sizeof(buf), "thread_%d", thid++);
+				if (procfs_pid_slurp (i, "comm", buf, sizeof(buf)) == -1) {
+					/* fall back to auto-id */
+					snprintf (buf, sizeof(buf), "thread_%d", thid++);
+				}
+
 				r_list_append (list, r_debug_pid_new (buf, i, 's', 0));
 			}
 		}
