+++ libr/debug/p/native/procfs.h
@@ -1,0 +1,3 @@
+/* radare - LGPL - Copyright 2009-2016 - pancake */
+
+int procfs_pid_slurp (int pid, char *prop, char *out, size_t len);
